--- src/app/b9-offer-real-money/page.tsx
@@ -1,33 +1,0 @@
-import { H4, H3, P } from '@/components/Typography'
-
-export default function OfferRealMoney(){
-    return(
-        <div>
-            <H3>B9 Game offers Real Money for Players?
-            </H3>
-            <P className='p-2'>B9Game Download is an official casino platform that works largely in Pakistan. These days, many people are earning a decent income through this game. It is a platform where you can access thousands of casino slots, cash machines, poker, blackjack, and other games. By playing these games, you can easily earn money from the comfort of your home. However, many users wonder, &quot;Does B9 Game offer real money for players?&quot; If you are also pondering this question, today we will confirm in our content whether this app truly provides real money or not.
-            </P>
-            <H3>Is B9 Game offers Real Money for Players?
-            </H3>
-            <P className='p-2'>If players are thinking that this platform is a scam or fraud, then that&apos;s absolutely not the case. This is an official casino app, and there is no fraud or scam involved. In this app, every user, whether they are a pro or a beginner, can benefit from the games and earn a substantial amount of money. Yes, you heard it right! The purpose of saying this is that B9 Game is a 100% real money-earning app that offers big wins and rewards to its users on a daily basis. And I can surely say that anyone can generate a decent income daily from this game and solve their financial issues from the comfort of their home.
-            </P>
-            <H3>Deposit Money & Withdrawals
-            </H3>
-            <H4>Deposit:
-            </H4>
-            <P className='p-2'>In this game, a player can deposit an amount ranging from 100 to 50,000. As a player, I would advise you to deposit a small amount according to your comfort and then select any of your favorite games available in the platform and start playing, while also enjoying the experience.
-            </P>
-            <H4>Withdrawal: 
-            </H4>
-            <P className='p-2'>Players can withdraw amounts ranging from 500 to 50,000 in a single day. If you want to withdraw your winnings after playing the game, simply click on the withdrawal option and transfer your winning amount to your Easypaisa, JazzCash, or bank account. After that, within a few hours you can receive your winning amount in your account.
-            </P>
-            <H3>B9 Game offers Real Money for Players?
-            </H3>
-            <P className='p-2'>B9 Game provides players with the opportunity to earn real money through engaging gameplay. With various games to choose from, players can enjoy a thrilling experience while potentially increasing their winnings. The platform allows withdrawals ranging from 500 to 50,000 in a single day, making it accessible for both casual and serious gamers. Players can easily withdraw their earnings to popular payment methods like Easypaisa, JazzCash, or directly to their bank accounts, ensuring a convenient and rewarding gaming experience.
-            </P>
-            <H3>End:
-            </H3>
-            <P className='p-2'>At the end, B9 Game stands out as an exciting platform for players seeking real money rewards. With a diverse range of games and the potential to win significant amounts, it caters to both novice and experienced gamers. The user-friendly interface and straightforward withdrawal process enhance the overall experience, allowing players to transfer their winnings easily to Easypaisa, JazzCash, or bank accounts. The daily withdrawal limit of 500 to 50,000 adds flexibility, making it accessible for all types of players. Overall, B9 Game offers a thrilling combination of entertainment and earning potential, making it a compelling choice for online gaming enthusiasts.</P>
-        </div>
-    )
-}
