--- src/app/disclaimer-page/page.tsx
@@ -1,37 +1,0 @@
-import { H3, H5, P } from '@/components/Typography'
-export default function DisclaimerPage(){
-    return(
-        <div>
-            <H3>Disclaimer for <a className='underline text-green-500' href="https://B9gameofficial.com" target="_blank" rel="noopener noreferrer">B9gameofficial.com</a>
-            </H3>
-            <H5>
-            Welcome to <a className='underline text-green-500' href="https://B9gameofficial.com" target="_blank" rel="noopener noreferrer">B9gameofficial.com</a>!
-            </H5>
-            <P className='p-2'>This disclaimer applies to your access and use of the B9 Game Official website (the &quot;Website&quot;). Please read this Disclaimer carefully before using the Website. By accessing or using the Website, you agree to be bound by this Disclaimer and its terms.   
-            </P>
-            <H3>Content Disclaimer
-            </H3>
-            <P className='p-2'>The content on the Website is for informational purposes only and should not be construed as professional advice. While we strive to provide accurate and up-to-date information, we make no guarantees or warranties about the accuracy, completeness, or reliability of any content on the Website.   
-            </P>
-            <H3>Links to Third-Party Websites
-            </H3>
-            <P className='p-2'>The Website may contain links to third-party websites. These links are provided for your convenience only and do not constitute an endorsement by us of the content on such websites. We are not responsible for the content of any linked websites.</P>
-            <H5>Intellectual Property
-            </H5>
-            <P className='p-2'>The content on the Website, including but not limited to text, graphics, logos, images, and software, is the property of <a className='underline text-green-500' href="https://B9gameofficial.com" target="_blank" rel="noopener noreferrer">B9gameofficial.com</a> or its licensors and is protected by copyright and other intellectual property laws. You may not reproduce, modify, distribute, or commercially exploit any content on the Website without our express written permission.</P>
-            <H3>Limitation of Liability
-            </H3>
-            <P className='p-2'>We will not be liable for any damages arising out of or related to your use of the Website, including but not limited to direct, indirect, incidental, consequential, or punitive damages.</P>
-            <H3>Governing Law
-            </H3>
-            <P className='p-2'>This Disclaimer shall be governed by and construed in accordance with the laws of [Pakistan]</P>
-            <H3>Changes to the Disclaimer
-            </H3>
-            <P className='p-2'>We reserve the right to change this Disclaimer at any time. The revised Disclaimer will be effective immediately upon posting on the Website.
-            </P>
-            <H3>Contact Us
-            </H3>
-            <P className='p-2'>If you have any questions about this Disclaimer, please contact us at [contact information for <a className='underline text-green-500' href="https://B9gameofficial.com" target="_blank" rel="noopener noreferrer">B9gameofficial.com</a>].</P>
-        </div>
-    )
-}
