--- src/app/real-or-fake-page/page.tsx
@@ -1,62 +1,0 @@
-import { H3, P } from '@/components/Typography'
-export default function RealOrFakePage(){
-    return(
-        <div>
-            <H3>B9 Game is Real or Fake?
-            </H3>
-            <P className='p-2'>Welcome to the app at <a className='underline text-green-500' href="https://B9gameofficial.com" target="_blank" rel="noopener noreferrer">B9gameofficial.com</a>. Today, we will discuss whether B9 Game is real or fake? This is a platform that has been trending in Pakistan recently. The application has become quite popular due to its realism and stunning graphics. However, despite this, many users still question whether the platform is fake or genuine. So, without wasting any time, let’s move forward and learn more about this amazing app.
-            </P>
-            <H3>Is B9 Game Real or Fake?
-            </H3>
-            <P className='p-2'>B9 Game is an earning application that provides both new and old users with a good opportunity to earn real money. I can assure you that if you place your bets strategically in this application, it will never give you a chance to lose. This is a 100% real earning application for all Android & iOS users. Furthermore, you can download it for free and generate a decent income on a daily basis without any major losses.
-            However, if you&apos;re still concerned about this application being fake, you can download it and play once or twice. It will give you the chance to win at every step, and I can confidently say that by downloading this app, you can consistently earn a rewarding income. Moreover, you can easily withdraw your earnings into your accounts. Do remember that you can only withdraw a limited amount from this game each day
-            </P>
-            <H3>Benefits in B9 Game for New Palyers
-            </H3>
-            <P className='p-2'>If you are a new user of this casino game and are downloading it for the first time, the game offers stunning gifts for you. As soon as a new user joins the game, they receive a certain amount of money as a bonus. Along with this, sign-up bonuses and weekly and monthly rewards are also provided for users. So, if you are a new user and you have doubts about losing in this game, don’t worry. This app offers many benefits for its new users. All you have to do is simply sign up, deposit a small amount, select your favorite game from the thousands available, and start generating a decent amount of income right from the first try.
-            </P>
-            <H3>List of Big Winning Games in B9 Game
-            </H3>
-            <P className='p-2'>Here is the list of some big winnings game in B9 Game Download:
-            </P>
-            <ol>
-                <li><P>Slots Machines
-                </P>
-                </li>
-                <li><P>Pokers
-                </P>
-                </li>
-                <li><P>Ludo
-                </P>
-                </li>
-                <li><P>Fishing Games
-                </P>
-                </li>
-                <li><P>Roulette
-                </P>
-                </li>
-                <li><P>Aviator
-                </P>
-                </li>
-                <li><P>Sports Betting
-                </P>
-                </li>
-                <li><P>Teen Patti
-                </P>
-                </li>
-                <li><P>Joker
-                </P>
-                </li>
-                <li><P>Blackjack
-                </P>
-                </li>
-                <li><P>And many more games are available in this platform.
-                </P></li>
-            </ol>
-            <H3>Download B9 Game Earn Money
-            </H3>
-            <P className='p-2'>As a user, want to earn money comfortably from home without any hassle, there is no need to work hard. Simply visit our website <a className='underline text-green-500' href="https://B9gameofficial.com" target="_blank" rel="noopener noreferrer">B9gameofficial.com</a> and download then install this amazing platform on your Android or iOS devices you have at no cost. After that, upon entering the game, you just need to deposit a small amount and start playing. Play the game in which you have skill and that is your favorite, so you can enjoy the game and generate a good income from it. So, what are you waiting for? Download this amazing game now and enjoy unlimited fun, while also taking advantage of this great opportunity to earn money from home.
-            </P>
-        </div>
-    )
-}
